move macro to std; add paste as a prelude
