+++ crates/std/Cargo.toml
@@ -15,6 +15,7 @@ grug-macros  = { path = "../macros" }
 grug-math    = { path = "../math" }
 grug-storage = { path = "../storage" }
 grug-types   = { path = "../types" }
+paste        = { workspace = true }
 serde        = { workspace = true, features = ["derive"] }
 serde_json   = { workspace = true }
 serde_with   = { workspace = true }
